Attach a Protocol to an Experiment on Labstep

This function allows you to attach a Protocol to an existing Experiment.

attachProtocol(user, experiment, protocol)

Arguments

user
A labstep user object. Must contain an `api_key` field. Returned from `login` command

experiment
The experiment you want to attach the protocol to. Must have at least an `id` field.

protocol
The protocol you want to attach. Must have at least an `id` field

Value

Returns an `experiment` object

Examples

user <- login("demo@labstep.com","demopassword")
expName = 'My First Experiment'
expDesc = 'An experiment testing the labstepR library'
experiment <- createExperiment(user,expName,expDesc)
protocol <- createProtocol(user,'My First Protocol')
experiment <- attachProtocol(user,experiment,protocol)
print(experiment)

#> $id
#> [1] 14686
#> 
#> $created_at
#> [1] "2019-04-01T16:51:40+00:00"
#> 
#> $updated_at
#> [1] "2019-04-01T16:51:40+00:00"
#> 
#> $deleted_at
#> NULL
#> 
#> $started_at
#> [1] "2019-04-01T16:51:40+00:00"
#> 
#> $ended_at
#> NULL
#> 
#> $name
#> [1] "My First Experiment"
#> 
#> $description
#> [1] "An experiment testing the labstepR library"
#> 
#> $author
#> $author$id
#> [1] 4524
#> 
#> $author$username
#> [1] "demo@labstep.com"
#> 
#> $author$name
#> [1] "Demo Labstep"
#> 
#> $author$first_name
#> [1] "Demo"
#> 
#> $author$last_name
#> [1] "Labstep"
#> 
#> $author$profile
#> $author$profile$id
#> [1] 5751
#> 
#> $author$profile$photo
#> NULL
#> 
#> 
#> 
#> $thread
#> $thread$comments_count
#> [1] 0
#> 
#> $thread$id
#> [1] 477132
#> 
#> 
#> $experiments
#> $experiments[[1]]
#> $experiments[[1]]$id
#> [1] 14299
#> 
#> $experiments[[1]]$created_at
#> [1] "2019-04-01T16:51:42+00:00"
#> 
#> $experiments[[1]]$deleted_at
#> NULL
#> 
#> $experiments[[1]]$started_at
#> [1] "2019-04-01T16:51:42+00:00"
#> 
#> $experiments[[1]]$ended_at
#> NULL
#> 
#> $experiments[[1]]$position
#> [1] 1
#> 
#> $experiments[[1]]$protocol
#> $experiments[[1]]$protocol$version
#> [1] 0
#> 
#> $experiments[[1]]$protocol$id
#> [1] 7480
#> 
#> $experiments[[1]]$protocol$created_at
#> [1] "2019-04-01T16:51:41+00:00"
#> 
#> $experiments[[1]]$protocol$name
#> [1] "My First Protocol"
#> 
#> $experiments[[1]]$protocol$collection
#> $experiments[[1]]$protocol$collection$id
#> [1] 6143
#> 
#> 
#> 
#> $experiments[[1]]$experiment_steps
#> $experiments[[1]]$experiment_steps[[1]]
#> $experiments[[1]]$experiment_steps[[1]]$id
#> [1] 50663
#> 
#> $experiments[[1]]$experiment_steps[[1]]$ended_at
#> NULL
#> 
#> $experiments[[1]]$experiment_steps[[1]]$thread
#> $experiments[[1]]$experiment_steps[[1]]$thread$comments_count
#> [1] 0
#> 
#> $experiments[[1]]$experiment_steps[[1]]$thread$id
#> [1] 477133
#> 
#> 
#> 
#> 
#> $experiments[[1]]$permissions
#> $experiments[[1]]$permissions$view
#> [1] TRUE
#> 
#> $experiments[[1]]$permissions$edit
#> [1] TRUE
#> 
#> $experiments[[1]]$permissions$owner
#> [1] TRUE
#> 
#> 
#> 
#> 
#> $signatures
#> list()
#> 
#> $permissions
#> $permissions$view
#> [1] TRUE
#> 
#> $permissions$edit
#> [1] TRUE
#> 
#> $permissions$owner
#> [1] TRUE
#> 
#>