Create a Protocol on Labstep

This function allows you to create a protocol.

createProtocol(user, name)

Arguments

user
A labstep user object. Must contain an `api_key` field. Returned from `login` command

name
The name you want to give your protocol.

Value

Returns an `protocol` object

Examples

user <- login("demo@labstep.com","demopassword")
protocol <- createProtocol(user,'My First Protocol')
print(protocol)

#> $all_experiment_count
#> [1] 0
#> 
#> $experiment_count
#> [1] 0
#> 
#> $protocol_step_count
#> [1] 1
#> 
#> $file_count
#> [1] 0
#> 
#> $protocol_table_count
#> [1] 0
#> 
#> $protocol_timer_count
#> [1] 0
#> 
#> $protocol_value_count
#> [1] 0
#> 
#> $children_protocol_collection_count
#> [1] 0
#> 
#> $version
#> [1] 0
#> 
#> $is_last_version
#> [1] TRUE
#> 
#> $id
#> [1] 7487
#> 
#> $created_at
#> [1] "2019-04-01T17:02:31+00:00"
#> 
#> $updated_at
#> [1] "2019-04-01T17:02:32+00:00"
#> 
#> $name
#> [1] "My First Protocol"
#> 
#> $changes
#> NULL
#> 
#> $additional_information
#> NULL
#> 
#> $content_state
#> $content_state$object
#> [1] "value"
#> 
#> $content_state$document
#> $content_state$document$data
#> list()
#> 
#> $content_state$document$nodes
#> $content_state$document$nodes[[1]]
#> $content_state$document$nodes[[1]]$object
#> [1] "block"
#> 
#> $content_state$document$nodes[[1]]$isVoid
#> [1] FALSE
#> 
#> $content_state$document$nodes[[1]]$type
#> [1] "heading"
#> 
#> $content_state$document$nodes[[1]]$data
#> list()
#> 
#> $content_state$document$nodes[[1]]$nodes
#> $content_state$document$nodes[[1]]$nodes[[1]]
#> $content_state$document$nodes[[1]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $content_state$document$nodes[[1]]$nodes[[1]]$leaves
#> $content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]
#> $content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Introduction"
#> 
#> $content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $content_state$document$nodes[[2]]
#> $content_state$document$nodes[[2]]$object
#> [1] "block"
#> 
#> $content_state$document$nodes[[2]]$type
#> [1] "paragraph"
#> 
#> $content_state$document$nodes[[2]]$data
#> list()
#> 
#> $content_state$document$nodes[[2]]$isVoid
#> [1] FALSE
#> 
#> $content_state$document$nodes[[2]]$nodes
#> $content_state$document$nodes[[2]]$nodes[[1]]
#> $content_state$document$nodes[[2]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $content_state$document$nodes[[2]]$nodes[[1]]$leaves
#> $content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]
#> $content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Get started by giving your protocol a name and by adding an introduction..."
#> 
#> $content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $content_state$document$nodes[[3]]
#> $content_state$document$nodes[[3]]$object
#> [1] "block"
#> 
#> $content_state$document$nodes[[3]]$isVoid
#> [1] FALSE
#> 
#> $content_state$document$nodes[[3]]$type
#> [1] "heading"
#> 
#> $content_state$document$nodes[[3]]$data
#> list()
#> 
#> $content_state$document$nodes[[3]]$nodes
#> $content_state$document$nodes[[3]]$nodes[[1]]
#> $content_state$document$nodes[[3]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $content_state$document$nodes[[3]]$nodes[[1]]$leaves
#> $content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]
#> $content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Materials"
#> 
#> $content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $content_state$document$nodes[[4]]
#> $content_state$document$nodes[[4]]$object
#> [1] "block"
#> 
#> $content_state$document$nodes[[4]]$type
#> [1] "paragraph"
#> 
#> $content_state$document$nodes[[4]]$data
#> list()
#> 
#> $content_state$document$nodes[[4]]$isVoid
#> [1] FALSE
#> 
#> $content_state$document$nodes[[4]]$nodes
#> $content_state$document$nodes[[4]]$nodes[[1]]
#> $content_state$document$nodes[[4]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $content_state$document$nodes[[4]]$nodes[[1]]$leaves
#> $content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]
#> $content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$text
#> [1] "You can add the resources (select the resource icon) you use in your protocol here or in the procedure."
#> 
#> $content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $content_state$document$nodes[[5]]
#> $content_state$document$nodes[[5]]$object
#> [1] "block"
#> 
#> $content_state$document$nodes[[5]]$isVoid
#> [1] FALSE
#> 
#> $content_state$document$nodes[[5]]$type
#> [1] "heading"
#> 
#> $content_state$document$nodes[[5]]$data
#> list()
#> 
#> $content_state$document$nodes[[5]]$nodes
#> $content_state$document$nodes[[5]]$nodes[[1]]
#> $content_state$document$nodes[[5]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $content_state$document$nodes[[5]]$nodes[[1]]$leaves
#> $content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]
#> $content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Procedure"
#> 
#> $content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $content_state$document$nodes[[6]]
#> $content_state$document$nodes[[6]]$object
#> [1] "block"
#> 
#> $content_state$document$nodes[[6]]$type
#> [1] "protocol_step"
#> 
#> $content_state$document$nodes[[6]]$isVoid
#> [1] FALSE
#> 
#> $content_state$document$nodes[[6]]$data
#> $content_state$document$nodes[[6]]$data$id
#> [1] 30138
#> 
#> 
#> $content_state$document$nodes[[6]]$nodes
#> $content_state$document$nodes[[6]]$nodes[[1]]
#> $content_state$document$nodes[[6]]$nodes[[1]]$object
#> [1] "block"
#> 
#> $content_state$document$nodes[[6]]$nodes[[1]]$type
#> [1] "paragraph"
#> 
#> $content_state$document$nodes[[6]]$nodes[[1]]$data
#> list()
#> 
#> $content_state$document$nodes[[6]]$nodes[[1]]$isVoid
#> [1] FALSE
#> 
#> $content_state$document$nodes[[6]]$nodes[[1]]$nodes
#> $content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]
#> $content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves
#> $content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]
#> $content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Detail your protocol procedure in steps like this. To add another step, click below 'Step 1' and select the plus icon on the left hand side."
#> 
#> $content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $content_state$document$object
#> [1] "document"
#> 
#> 
#> 
#> $author
#> $author$id
#> [1] 4524
#> 
#> $author$username
#> [1] "demo@labstep.com"
#> 
#> $author$roles
#> $author$roles[[1]]
#> [1] "ROLE_USER"
#> 
#> $author$roles[[2]]
#> [1] "ROLE_GROUP_1182"
#> 
#> $author$roles[[3]]
#> [1] "ROLE_GROUP_OWNER_1182"
#> 
#> $author$roles[[4]]
#> [1] "ROLE_GROUP_1183"
#> 
#> $author$roles[[5]]
#> [1] "ROLE_GROUP_OWNER_1183"
#> 
#> $author$roles[[6]]
#> [1] "ROLE_GROUP_1184"
#> 
#> $author$roles[[7]]
#> [1] "ROLE_GROUP_OWNER_1184"
#> 
#> $author$roles[[8]]
#> [1] "ROLE_GROUP_1185"
#> 
#> $author$roles[[9]]
#> [1] "ROLE_GROUP_OWNER_1185"
#> 
#> $author$roles[[10]]
#> [1] "ROLE_GROUP_1186"
#> 
#> $author$roles[[11]]
#> [1] "ROLE_GROUP_OWNER_1186"
#> 
#> $author$roles[[12]]
#> [1] "ROLE_GROUP_1187"
#> 
#> $author$roles[[13]]
#> [1] "ROLE_GROUP_OWNER_1187"
#> 
#> $author$roles[[14]]
#> [1] "ROLE_GROUP_1188"
#> 
#> $author$roles[[15]]
#> [1] "ROLE_GROUP_OWNER_1188"
#> 
#> $author$roles[[16]]
#> [1] "ROLE_GROUP_1189"
#> 
#> $author$roles[[17]]
#> [1] "ROLE_GROUP_OWNER_1189"
#> 
#> $author$roles[[18]]
#> [1] "ROLE_GROUP_1190"
#> 
#> $author$roles[[19]]
#> [1] "ROLE_GROUP_OWNER_1190"
#> 
#> $author$roles[[20]]
#> [1] "ROLE_GROUP_1191"
#> 
#> $author$roles[[21]]
#> [1] "ROLE_GROUP_OWNER_1191"
#> 
#> $author$roles[[22]]
#> [1] "ROLE_GROUP_1192"
#> 
#> $author$roles[[23]]
#> [1] "ROLE_GROUP_OWNER_1192"
#> 
#> $author$roles[[24]]
#> [1] "ROLE_GROUP_1193"
#> 
#> $author$roles[[25]]
#> [1] "ROLE_GROUP_OWNER_1193"
#> 
#> $author$roles[[26]]
#> [1] "ROLE_GROUP_1198"
#> 
#> $author$roles[[27]]
#> [1] "ROLE_GROUP_OWNER_1198"
#> 
#> $author$roles[[28]]
#> [1] "ROLE_GROUP_1200"
#> 
#> $author$roles[[29]]
#> [1] "ROLE_GROUP_OWNER_1200"
#> 
#> $author$roles[[30]]
#> [1] "ROLE_GROUP_1201"
#> 
#> $author$roles[[31]]
#> [1] "ROLE_GROUP_OWNER_1201"
#> 
#> $author$roles[[32]]
#> [1] "ROLE_GROUP_1202"
#> 
#> $author$roles[[33]]
#> [1] "ROLE_GROUP_OWNER_1202"
#> 
#> $author$roles[[34]]
#> [1] "ROLE_GROUP_1203"
#> 
#> $author$roles[[35]]
#> [1] "ROLE_GROUP_OWNER_1203"
#> 
#> $author$roles[[36]]
#> [1] "ROLE_GROUP_1204"
#> 
#> $author$roles[[37]]
#> [1] "ROLE_GROUP_OWNER_1204"
#> 
#> $author$roles[[38]]
#> [1] "ROLE_GROUP_1205"
#> 
#> $author$roles[[39]]
#> [1] "ROLE_GROUP_OWNER_1205"
#> 
#> $author$roles[[40]]
#> [1] "ROLE_GROUP_1206"
#> 
#> $author$roles[[41]]
#> [1] "ROLE_GROUP_OWNER_1206"
#> 
#> $author$roles[[42]]
#> [1] "ROLE_GROUP_1207"
#> 
#> $author$roles[[43]]
#> [1] "ROLE_GROUP_OWNER_1207"
#> 
#> $author$roles[[44]]
#> [1] "ROLE_GROUP_1208"
#> 
#> $author$roles[[45]]
#> [1] "ROLE_GROUP_OWNER_1208"
#> 
#> 
#> $author$name
#> [1] "Demo Labstep"
#> 
#> $author$first_name
#> [1] "Demo"
#> 
#> $author$last_name
#> [1] "Labstep"
#> 
#> 
#> $collection
#> $collection$id
#> [1] 6150
#> 
#> $collection$created_at
#> [1] "2019-04-01T17:02:31+00:00"
#> 
#> $collection$updated_at
#> [1] "2019-04-01T17:02:31+00:00"
#> 
#> $collection$deleted_at
#> NULL
#> 
#> $collection$parent_protocol
#> NULL
#> 
#> $collection$logged_user_protocol_collection_playlist_count
#> [1] 1
#> 
#> $collection$tags
#> list()
#> 
#> $collection$permissions
#> $collection$permissions$view
#> [1] TRUE
#> 
#> $collection$permissions$edit
#> [1] TRUE
#> 
#> $collection$permissions$owner
#> [1] TRUE
#> 
#> 
#> 
#> $protocol_collection_draft
#> NULL
#> 
#> $logged_user_experiment_count
#> [1] 0
#> 
#> $children_protocol_collections
#> list()
#> 
#> $protocol_steps
#> $protocol_steps[[1]]
#> $protocol_steps[[1]]$id
#> [1] 30138
#> 
#> $protocol_steps[[1]]$files
#> list()
#> 
#> $protocol_steps[[1]]$protocol_tables
#> list()
#> 
#> $protocol_steps[[1]]$protocol_timers
#> list()
#> 
#> $protocol_steps[[1]]$protocol_values
#> list()
#> 
#> 
#> 
#> $protocol_tables
#> list()
#> 
#> $protocol_timers
#> list()
#> 
#> $protocol_values
#> list()
#> 
#> $files
#> list()
#>