Get Protocol

This function allows you to retrieve metadata for a single protocol

getProtocol(user, id)

Arguments

user
A labstep user object. Must contain an `api_key` field. Returned from `login` command

id
ID of the protocol to retrieve

Value

Returns a `protocol` object

Examples

user <- login("demo@labstep.com","demopassword")
protocol <- getProtocol(user,6140)
print(protocol)

#> $last_version
#> $last_version$all_experiment_count
#> [1] 0
#> 
#> $last_version$experiment_count
#> [1] 0
#> 
#> $last_version$protocol_step_count
#> [1] 1
#> 
#> $last_version$protocol_table_count
#> [1] 0
#> 
#> $last_version$protocol_timer_count
#> [1] 0
#> 
#> $last_version$protocol_value_count
#> [1] 0
#> 
#> $last_version$children_protocol_collection_count
#> [1] 0
#> 
#> $last_version$version
#> [1] 0
#> 
#> $last_version$is_last_version
#> [1] TRUE
#> 
#> $last_version$id
#> [1] 7477
#> 
#> $last_version$name
#> [1] "My First Protocol"
#> 
#> $last_version$changes
#> NULL
#> 
#> $last_version$additional_information
#> NULL
#> 
#> $last_version$content_state
#> $last_version$content_state$object
#> [1] "value"
#> 
#> $last_version$content_state$document
#> $last_version$content_state$document$data
#> list()
#> 
#> $last_version$content_state$document$nodes
#> $last_version$content_state$document$nodes[[1]]
#> $last_version$content_state$document$nodes[[1]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[1]]$type
#> [1] "heading"
#> 
#> $last_version$content_state$document$nodes[[1]]$nodes
#> $last_version$content_state$document$nodes[[1]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[1]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Introduction"
#> 
#> $last_version$content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $last_version$content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[1]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[1]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[1]]$object
#> [1] "block"
#> 
#> 
#> $last_version$content_state$document$nodes[[2]]
#> $last_version$content_state$document$nodes[[2]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[2]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[2]]$nodes
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Get started by giving your protocol a name and by adding an introduction..."
#> 
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[2]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[2]]$object
#> [1] "block"
#> 
#> 
#> $last_version$content_state$document$nodes[[3]]
#> $last_version$content_state$document$nodes[[3]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[3]]$type
#> [1] "heading"
#> 
#> $last_version$content_state$document$nodes[[3]]$nodes
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Materials"
#> 
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[3]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[3]]$object
#> [1] "block"
#> 
#> 
#> $last_version$content_state$document$nodes[[4]]
#> $last_version$content_state$document$nodes[[4]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[4]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[4]]$nodes
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$text
#> [1] "You can add the resources (select the resource icon) you use in your protocol here or in the procedure."
#> 
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[4]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[4]]$object
#> [1] "block"
#> 
#> 
#> $last_version$content_state$document$nodes[[5]]
#> $last_version$content_state$document$nodes[[5]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[5]]$type
#> [1] "heading"
#> 
#> $last_version$content_state$document$nodes[[5]]$nodes
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Procedure"
#> 
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[5]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[5]]$object
#> [1] "block"
#> 
#> 
#> $last_version$content_state$document$nodes[[6]]
#> $last_version$content_state$document$nodes[[6]]$data
#> $last_version$content_state$document$nodes[[6]]$data$id
#> [1] 30128
#> 
#> 
#> $last_version$content_state$document$nodes[[6]]$type
#> [1] "protocol_step"
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$nodes
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Detail your protocol procedure in steps like this. To add another step, click below 'Step 1' and select the plus icon on the left hand side."
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$object
#> [1] "block"
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[6]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[6]]$object
#> [1] "block"
#> 
#> 
#> 
#> $last_version$content_state$document$object
#> [1] "document"
#> 
#> 
#> 
#> $last_version$protocol_steps
#> $last_version$protocol_steps[[1]]
#> $last_version$protocol_steps[[1]]$id
#> [1] 30128
#> 
#> $last_version$protocol_steps[[1]]$files
#> list()
#> 
#> $last_version$protocol_steps[[1]]$protocol_tables
#> list()
#> 
#> $last_version$protocol_steps[[1]]$protocol_timers
#> list()
#> 
#> $last_version$protocol_steps[[1]]$protocol_values
#> list()
#> 
#> 
#> 
#> $last_version$protocol_tables
#> list()
#> 
#> $last_version$protocol_timers
#> list()
#> 
#> $last_version$protocol_values
#> list()
#> 
#> $last_version$files
#> list()
#> 
#> 
#> $id
#> [1] 6140
#> 
#> $created_at
#> [1] "2019-04-01T16:47:50+00:00"
#> 
#> $updated_at
#> [1] "2019-04-01T16:47:50+00:00"
#> 
#> $deleted_at
#> NULL
#> 
#> $author
#> $author$id
#> [1] 4524
#> 
#> $author$username
#> [1] "demo@labstep.com"
#> 
#> $author$profile
#> $author$profile$id
#> [1] 5751
#> 
#> $author$profile$photo
#> NULL
#> 
#> 
#> 
#> $draft
#> NULL
#> 
#> $parent_protocol
#> NULL
#> 
#> $versions
#> $versions[[1]]
#> $versions[[1]]$all_experiment_count
#> [1] 0
#> 
#> $versions[[1]]$experiment_count
#> [1] 0
#> 
#> $versions[[1]]$protocol_step_count
#> [1] 1
#> 
#> $versions[[1]]$protocol_table_count
#> [1] 0
#> 
#> $versions[[1]]$protocol_timer_count
#> [1] 0
#> 
#> $versions[[1]]$protocol_value_count
#> [1] 0
#> 
#> $versions[[1]]$children_protocol_collection_count
#> [1] 0
#> 
#> $versions[[1]]$version
#> [1] 0
#> 
#> $versions[[1]]$is_last_version
#> [1] TRUE
#> 
#> $versions[[1]]$id
#> [1] 7477
#> 
#> $versions[[1]]$name
#> [1] "My First Protocol"
#> 
#> $versions[[1]]$changes
#> NULL
#> 
#> $versions[[1]]$additional_information
#> NULL
#> 
#> $versions[[1]]$content_state
#> $versions[[1]]$content_state$object
#> [1] "value"
#> 
#> $versions[[1]]$content_state$document
#> $versions[[1]]$content_state$document$data
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes
#> $versions[[1]]$content_state$document$nodes[[1]]
#> $versions[[1]]$content_state$document$nodes[[1]]$data
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[1]]$type
#> [1] "heading"
#> 
#> $versions[[1]]$content_state$document$nodes[[1]]$nodes
#> $versions[[1]]$content_state$document$nodes[[1]]$nodes[[1]]
#> $versions[[1]]$content_state$document$nodes[[1]]$nodes[[1]]$leaves
#> $versions[[1]]$content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]
#> $versions[[1]]$content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Introduction"
#> 
#> $versions[[1]]$content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[1]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[1]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[1]]$isVoid
#> [1] FALSE
#> 
#> $versions[[1]]$content_state$document$nodes[[1]]$object
#> [1] "block"
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[2]]
#> $versions[[1]]$content_state$document$nodes[[2]]$data
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[2]]$type
#> [1] "paragraph"
#> 
#> $versions[[1]]$content_state$document$nodes[[2]]$nodes
#> $versions[[1]]$content_state$document$nodes[[2]]$nodes[[1]]
#> $versions[[1]]$content_state$document$nodes[[2]]$nodes[[1]]$leaves
#> $versions[[1]]$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]
#> $versions[[1]]$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Get started by giving your protocol a name and by adding an introduction..."
#> 
#> $versions[[1]]$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[2]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[2]]$isVoid
#> [1] FALSE
#> 
#> $versions[[1]]$content_state$document$nodes[[2]]$object
#> [1] "block"
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[3]]
#> $versions[[1]]$content_state$document$nodes[[3]]$data
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[3]]$type
#> [1] "heading"
#> 
#> $versions[[1]]$content_state$document$nodes[[3]]$nodes
#> $versions[[1]]$content_state$document$nodes[[3]]$nodes[[1]]
#> $versions[[1]]$content_state$document$nodes[[3]]$nodes[[1]]$leaves
#> $versions[[1]]$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]
#> $versions[[1]]$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Materials"
#> 
#> $versions[[1]]$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[3]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[3]]$isVoid
#> [1] FALSE
#> 
#> $versions[[1]]$content_state$document$nodes[[3]]$object
#> [1] "block"
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[4]]
#> $versions[[1]]$content_state$document$nodes[[4]]$data
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[4]]$type
#> [1] "paragraph"
#> 
#> $versions[[1]]$content_state$document$nodes[[4]]$nodes
#> $versions[[1]]$content_state$document$nodes[[4]]$nodes[[1]]
#> $versions[[1]]$content_state$document$nodes[[4]]$nodes[[1]]$leaves
#> $versions[[1]]$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]
#> $versions[[1]]$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$text
#> [1] "You can add the resources (select the resource icon) you use in your protocol here or in the procedure."
#> 
#> $versions[[1]]$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[4]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[4]]$isVoid
#> [1] FALSE
#> 
#> $versions[[1]]$content_state$document$nodes[[4]]$object
#> [1] "block"
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[5]]
#> $versions[[1]]$content_state$document$nodes[[5]]$data
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[5]]$type
#> [1] "heading"
#> 
#> $versions[[1]]$content_state$document$nodes[[5]]$nodes
#> $versions[[1]]$content_state$document$nodes[[5]]$nodes[[1]]
#> $versions[[1]]$content_state$document$nodes[[5]]$nodes[[1]]$leaves
#> $versions[[1]]$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]
#> $versions[[1]]$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Procedure"
#> 
#> $versions[[1]]$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[5]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[5]]$isVoid
#> [1] FALSE
#> 
#> $versions[[1]]$content_state$document$nodes[[5]]$object
#> [1] "block"
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]
#> $versions[[1]]$content_state$document$nodes[[6]]$data
#> $versions[[1]]$content_state$document$nodes[[6]]$data$id
#> [1] 30128
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$type
#> [1] "protocol_step"
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$data
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$type
#> [1] "paragraph"
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$nodes
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Detail your protocol procedure in steps like this. To add another step, click below 'Step 1' and select the plus icon on the left hand side."
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$nodes[[1]]$object
#> [1] "text"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$isVoid
#> [1] FALSE
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$nodes[[1]]$object
#> [1] "block"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$isVoid
#> [1] FALSE
#> 
#> $versions[[1]]$content_state$document$nodes[[6]]$object
#> [1] "block"
#> 
#> 
#> 
#> $versions[[1]]$content_state$document$object
#> [1] "document"
#> 
#> 
#> 
#> $versions[[1]]$protocol_steps
#> $versions[[1]]$protocol_steps[[1]]
#> $versions[[1]]$protocol_steps[[1]]$id
#> [1] 30128
#> 
#> $versions[[1]]$protocol_steps[[1]]$files
#> list()
#> 
#> $versions[[1]]$protocol_steps[[1]]$protocol_tables
#> list()
#> 
#> $versions[[1]]$protocol_steps[[1]]$protocol_timers
#> list()
#> 
#> $versions[[1]]$protocol_steps[[1]]$protocol_values
#> list()
#> 
#> 
#> 
#> $versions[[1]]$protocol_tables
#> list()
#> 
#> $versions[[1]]$protocol_timers
#> list()
#> 
#> $versions[[1]]$protocol_values
#> list()
#> 
#> $versions[[1]]$files
#> list()
#> 
#> 
#> 
#> $logged_user_protocol_collection_playlist_count
#> [1] 1
#> 
#> $permissions
#> $permissions$view
#> [1] TRUE
#> 
#> $permissions$edit
#> [1] TRUE
#> 
#> $permissions$owner
#> [1] TRUE
#> 
#>