Authenticate a Labstep user

This function allows you to retrieve the user details for your Labstep account. It also returns your unique API key which you need to use the other Labstep API functions.

login(username, password)

Arguments

username
Your Labstep username (usually the email address you signed up with)

password
Your Labstep password

Value

Returns a 'user' object.

Examples

user <- login("demo@labstep.com","demopassword")
print(user)

#> $id
#> [1] 4524
#> 
#> $username
#> [1] "demo@labstep.com"
#> 
#> $enabled
#> [1] TRUE
#> 
#> $groups
#> $groups[[1]]
#> $groups[[1]]$id
#> [1] 1182
#> 
#> 
#> $groups[[2]]
#> $groups[[2]]$id
#> [1] 1183
#> 
#> 
#> $groups[[3]]
#> $groups[[3]]$id
#> [1] 1184
#> 
#> 
#> $groups[[4]]
#> $groups[[4]]$id
#> [1] 1185
#> 
#> 
#> $groups[[5]]
#> $groups[[5]]$id
#> [1] 1186
#> 
#> 
#> $groups[[6]]
#> $groups[[6]]$id
#> [1] 1187
#> 
#> 
#> $groups[[7]]
#> $groups[[7]]$id
#> [1] 1188
#> 
#> 
#> $groups[[8]]
#> $groups[[8]]$id
#> [1] 1189
#> 
#> 
#> $groups[[9]]
#> $groups[[9]]$id
#> [1] 1190
#> 
#> 
#> $groups[[10]]
#> $groups[[10]]$id
#> [1] 1191
#> 
#> 
#> $groups[[11]]
#> $groups[[11]]$id
#> [1] 1192
#> 
#> 
#> $groups[[12]]
#> $groups[[12]]$id
#> [1] 1193
#> 
#> 
#> $groups[[13]]
#> $groups[[13]]$id
#> [1] 1198
#> 
#> 
#> $groups[[14]]
#> $groups[[14]]$id
#> [1] 1200
#> 
#> 
#> $groups[[15]]
#> $groups[[15]]$id
#> [1] 1201
#> 
#> 
#> $groups[[16]]
#> $groups[[16]]$id
#> [1] 1202
#> 
#> 
#> $groups[[17]]
#> $groups[[17]]$id
#> [1] 1203
#> 
#> 
#> $groups[[18]]
#> $groups[[18]]$id
#> [1] 1204
#> 
#> 
#> $groups[[19]]
#> $groups[[19]]$id
#> [1] 1205
#> 
#> 
#> $groups[[20]]
#> $groups[[20]]$id
#> [1] 1206
#> 
#> 
#> $groups[[21]]
#> $groups[[21]]$id
#> [1] 1207
#> 
#> 
#> $groups[[22]]
#> $groups[[22]]$id
#> [1] 1208
#> 
#> 
#> $groups[[23]]
#> $groups[[23]]$id
#> [1] 1213
#> 
#> 
#> $groups[[24]]
#> $groups[[24]]$id
#> [1] 1214
#> 
#> 
#> $groups[[25]]
#> $groups[[25]]$id
#> [1] 1220
#> 
#> 
#> $groups[[26]]
#> $groups[[26]]$id
#> [1] 1221
#> 
#> 
#> $groups[[27]]
#> $groups[[27]]$id
#> [1] 1222
#> 
#> 
#> 
#> $roles
#> $roles[[1]]
#> [1] "ROLE_USER"
#> 
#> $roles[[2]]
#> [1] "ROLE_GROUP_1182"
#> 
#> $roles[[3]]
#> [1] "ROLE_GROUP_OWNER_1182"
#> 
#> $roles[[4]]
#> [1] "ROLE_GROUP_1183"
#> 
#> $roles[[5]]
#> [1] "ROLE_GROUP_OWNER_1183"
#> 
#> $roles[[6]]
#> [1] "ROLE_GROUP_1184"
#> 
#> $roles[[7]]
#> [1] "ROLE_GROUP_OWNER_1184"
#> 
#> $roles[[8]]
#> [1] "ROLE_GROUP_1185"
#> 
#> $roles[[9]]
#> [1] "ROLE_GROUP_OWNER_1185"
#> 
#> $roles[[10]]
#> [1] "ROLE_GROUP_1186"
#> 
#> $roles[[11]]
#> [1] "ROLE_GROUP_OWNER_1186"
#> 
#> $roles[[12]]
#> [1] "ROLE_GROUP_1187"
#> 
#> $roles[[13]]
#> [1] "ROLE_GROUP_OWNER_1187"
#> 
#> $roles[[14]]
#> [1] "ROLE_GROUP_1188"
#> 
#> $roles[[15]]
#> [1] "ROLE_GROUP_OWNER_1188"
#> 
#> $roles[[16]]
#> [1] "ROLE_GROUP_1189"
#> 
#> $roles[[17]]
#> [1] "ROLE_GROUP_OWNER_1189"
#> 
#> $roles[[18]]
#> [1] "ROLE_GROUP_1190"
#> 
#> $roles[[19]]
#> [1] "ROLE_GROUP_OWNER_1190"
#> 
#> $roles[[20]]
#> [1] "ROLE_GROUP_1191"
#> 
#> $roles[[21]]
#> [1] "ROLE_GROUP_OWNER_1191"
#> 
#> $roles[[22]]
#> [1] "ROLE_GROUP_1192"
#> 
#> $roles[[23]]
#> [1] "ROLE_GROUP_OWNER_1192"
#> 
#> $roles[[24]]
#> [1] "ROLE_GROUP_1193"
#> 
#> $roles[[25]]
#> [1] "ROLE_GROUP_OWNER_1193"
#> 
#> $roles[[26]]
#> [1] "ROLE_GROUP_1198"
#> 
#> $roles[[27]]
#> [1] "ROLE_GROUP_OWNER_1198"
#> 
#> $roles[[28]]
#> [1] "ROLE_GROUP_1200"
#> 
#> $roles[[29]]
#> [1] "ROLE_GROUP_OWNER_1200"
#> 
#> $roles[[30]]
#> [1] "ROLE_GROUP_1201"
#> 
#> $roles[[31]]
#> [1] "ROLE_GROUP_OWNER_1201"
#> 
#> $roles[[32]]
#> [1] "ROLE_GROUP_1202"
#> 
#> $roles[[33]]
#> [1] "ROLE_GROUP_OWNER_1202"
#> 
#> $roles[[34]]
#> [1] "ROLE_GROUP_1203"
#> 
#> $roles[[35]]
#> [1] "ROLE_GROUP_OWNER_1203"
#> 
#> $roles[[36]]
#> [1] "ROLE_GROUP_1204"
#> 
#> $roles[[37]]
#> [1] "ROLE_GROUP_OWNER_1204"
#> 
#> $roles[[38]]
#> [1] "ROLE_GROUP_1205"
#> 
#> $roles[[39]]
#> [1] "ROLE_GROUP_OWNER_1205"
#> 
#> $roles[[40]]
#> [1] "ROLE_GROUP_1206"
#> 
#> $roles[[41]]
#> [1] "ROLE_GROUP_OWNER_1206"
#> 
#> $roles[[42]]
#> [1] "ROLE_GROUP_1207"
#> 
#> $roles[[43]]
#> [1] "ROLE_GROUP_OWNER_1207"
#> 
#> $roles[[44]]
#> [1] "ROLE_GROUP_1208"
#> 
#> $roles[[45]]
#> [1] "ROLE_GROUP_OWNER_1208"
#> 
#> $roles[[46]]
#> [1] "ROLE_GROUP_1213"
#> 
#> $roles[[47]]
#> [1] "ROLE_GROUP_OWNER_1213"
#> 
#> $roles[[48]]
#> [1] "ROLE_GROUP_1214"
#> 
#> $roles[[49]]
#> [1] "ROLE_GROUP_OWNER_1214"
#> 
#> $roles[[50]]
#> [1] "ROLE_GROUP_1220"
#> 
#> $roles[[51]]
#> [1] "ROLE_GROUP_OWNER_1220"
#> 
#> $roles[[52]]
#> [1] "ROLE_GROUP_1221"
#> 
#> $roles[[53]]
#> [1] "ROLE_GROUP_OWNER_1221"
#> 
#> $roles[[54]]
#> [1] "ROLE_GROUP_1222"
#> 
#> $roles[[55]]
#> [1] "ROLE_GROUP_OWNER_1222"
#> 
#> 
#> $group_count
#> [1] 27
#> 
#> $name
#> [1] "Demo Labstep"
#> 
#> $authenticated
#> NULL
#> 
#> $auth_provider
#> NULL
#> 
#> $created_at
#> [1] "2019-03-25T17:08:28+00:00"
#> 
#> $consented_at
#> NULL
#> 
#> $first_name
#> [1] "Demo"
#> 
#> $last_name
#> [1] "Labstep"
#> 
#> $api_key
#> [1] "NmQxZjUzNTZlYzZiZDk0M2NmZTc2YzBlOGY3NGM2ZDdlZWM0NWMzMg=="
#> 
#> $experience
#> [1] 0
#> 
#> $level
#> [1] 1
#> 
#> $profile
#> $profile$id
#> [1] 5751
#> 
#> $profile$about
#> NULL
#> 
#> $profile$photo
#> NULL
#> 
#> $profile$university
#> NULL
#> 
#> $profile$papers
#> list()
#> 
#> 
#> $protocol_collection_playlists
#> $protocol_collection_playlists[[1]]
#> $protocol_collection_playlists[[1]]$id
#> [1] 4431
#> 
#> $protocol_collection_playlists[[1]]$name
#> [1] "My protocol playlist"
#> 
#> 
#> 
#> $resource_lists
#> $resource_lists[[1]]
#> $resource_lists[[1]]$id
#> [1] 4431
#> 
#> 
#> 
#> $user_groups
#> $user_groups[[1]]
#> $user_groups[[1]]$id
#> [1] 1969
#> 
#> $user_groups[[1]]$type
#> [1] "owner"
#> 
#> $user_groups[[1]]$security_policy
#> NULL
#> 
#> $user_groups[[1]]$group
#> $user_groups[[1]]$group$id
#> [1] 1182
#> 
#> 
#> 
#> $user_groups[[2]]
#> $user_groups[[2]]$id
#> [1] 1970
#> 
#> $user_groups[[2]]$type
#> [1] "owner"
#> 
#> $user_groups[[2]]$security_policy
#> NULL
#> 
#> $user_groups[[2]]$group
#> $user_groups[[2]]$group$id
#> [1] 1183
#> 
#> 
#> 
#> $user_groups[[3]]
#> $user_groups[[3]]$id
#> [1] 1971
#> 
#> $user_groups[[3]]$type
#> [1] "owner"
#> 
#> $user_groups[[3]]$security_policy
#> NULL
#> 
#> $user_groups[[3]]$group
#> $user_groups[[3]]$group$id
#> [1] 1184
#> 
#> 
#> 
#> $user_groups[[4]]
#> $user_groups[[4]]$id
#> [1] 1972
#> 
#> $user_groups[[4]]$type
#> [1] "owner"
#> 
#> $user_groups[[4]]$security_policy
#> NULL
#> 
#> $user_groups[[4]]$group
#> $user_groups[[4]]$group$id
#> [1] 1185
#> 
#> 
#> 
#> $user_groups[[5]]
#> $user_groups[[5]]$id
#> [1] 1973
#> 
#> $user_groups[[5]]$type
#> [1] "owner"
#> 
#> $user_groups[[5]]$security_policy
#> NULL
#> 
#> $user_groups[[5]]$group
#> $user_groups[[5]]$group$id
#> [1] 1186
#> 
#> 
#> 
#> $user_groups[[6]]
#> $user_groups[[6]]$id
#> [1] 1974
#> 
#> $user_groups[[6]]$type
#> [1] "owner"
#> 
#> $user_groups[[6]]$security_policy
#> NULL
#> 
#> $user_groups[[6]]$group
#> $user_groups[[6]]$group$id
#> [1] 1187
#> 
#> 
#> 
#> $user_groups[[7]]
#> $user_groups[[7]]$id
#> [1] 1975
#> 
#> $user_groups[[7]]$type
#> [1] "owner"
#> 
#> $user_groups[[7]]$security_policy
#> NULL
#> 
#> $user_groups[[7]]$group
#> $user_groups[[7]]$group$id
#> [1] 1188
#> 
#> 
#> 
#> $user_groups[[8]]
#> $user_groups[[8]]$id
#> [1] 1976
#> 
#> $user_groups[[8]]$type
#> [1] "owner"
#> 
#> $user_groups[[8]]$security_policy
#> NULL
#> 
#> $user_groups[[8]]$group
#> $user_groups[[8]]$group$id
#> [1] 1189
#> 
#> 
#> 
#> $user_groups[[9]]
#> $user_groups[[9]]$id
#> [1] 1977
#> 
#> $user_groups[[9]]$type
#> [1] "owner"
#> 
#> $user_groups[[9]]$security_policy
#> NULL
#> 
#> $user_groups[[9]]$group
#> $user_groups[[9]]$group$id
#> [1] 1190
#> 
#> 
#> 
#> $user_groups[[10]]
#> $user_groups[[10]]$id
#> [1] 1978
#> 
#> $user_groups[[10]]$type
#> [1] "owner"
#> 
#> $user_groups[[10]]$security_policy
#> NULL
#> 
#> $user_groups[[10]]$group
#> $user_groups[[10]]$group$id
#> [1] 1191
#> 
#> 
#> 
#> $user_groups[[11]]
#> $user_groups[[11]]$id
#> [1] 1979
#> 
#> $user_groups[[11]]$type
#> [1] "owner"
#> 
#> $user_groups[[11]]$security_policy
#> NULL
#> 
#> $user_groups[[11]]$group
#> $user_groups[[11]]$group$id
#> [1] 1192
#> 
#> 
#> 
#> $user_groups[[12]]
#> $user_groups[[12]]$id
#> [1] 1980
#> 
#> $user_groups[[12]]$type
#> [1] "owner"
#> 
#> $user_groups[[12]]$security_policy
#> NULL
#> 
#> $user_groups[[12]]$group
#> $user_groups[[12]]$group$id
#> [1] 1193
#> 
#> 
#> 
#> $user_groups[[13]]
#> $user_groups[[13]]$id
#> [1] 1985
#> 
#> $user_groups[[13]]$type
#> [1] "owner"
#> 
#> $user_groups[[13]]$security_policy
#> NULL
#> 
#> $user_groups[[13]]$group
#> $user_groups[[13]]$group$id
#> [1] 1198
#> 
#> 
#> 
#> $user_groups[[14]]
#> $user_groups[[14]]$id
#> [1] 1987
#> 
#> $user_groups[[14]]$type
#> [1] "owner"
#> 
#> $user_groups[[14]]$security_policy
#> NULL
#> 
#> $user_groups[[14]]$group
#> $user_groups[[14]]$group$id
#> [1] 1200
#> 
#> 
#> 
#> $user_groups[[15]]
#> $user_groups[[15]]$id
#> [1] 1988
#> 
#> $user_groups[[15]]$type
#> [1] "owner"
#> 
#> $user_groups[[15]]$security_policy
#> NULL
#> 
#> $user_groups[[15]]$group
#> $user_groups[[15]]$group$id
#> [1] 1201
#> 
#> 
#> 
#> $user_groups[[16]]
#> $user_groups[[16]]$id
#> [1] 1989
#> 
#> $user_groups[[16]]$type
#> [1] "owner"
#> 
#> $user_groups[[16]]$security_policy
#> NULL
#> 
#> $user_groups[[16]]$group
#> $user_groups[[16]]$group$id
#> [1] 1202
#> 
#> 
#> 
#> $user_groups[[17]]
#> $user_groups[[17]]$id
#> [1] 1990
#> 
#> $user_groups[[17]]$type
#> [1] "owner"
#> 
#> $user_groups[[17]]$security_policy
#> NULL
#> 
#> $user_groups[[17]]$group
#> $user_groups[[17]]$group$id
#> [1] 1203
#> 
#> 
#> 
#> $user_groups[[18]]
#> $user_groups[[18]]$id
#> [1] 1991
#> 
#> $user_groups[[18]]$type
#> [1] "owner"
#> 
#> $user_groups[[18]]$security_policy
#> NULL
#> 
#> $user_groups[[18]]$group
#> $user_groups[[18]]$group$id
#> [1] 1204
#> 
#> 
#> 
#> $user_groups[[19]]
#> $user_groups[[19]]$id
#> [1] 1992
#> 
#> $user_groups[[19]]$type
#> [1] "owner"
#> 
#> $user_groups[[19]]$security_policy
#> NULL
#> 
#> $user_groups[[19]]$group
#> $user_groups[[19]]$group$id
#> [1] 1205
#> 
#> 
#> 
#> $user_groups[[20]]
#> $user_groups[[20]]$id
#> [1] 1993
#> 
#> $user_groups[[20]]$type
#> [1] "owner"
#> 
#> $user_groups[[20]]$security_policy
#> NULL
#> 
#> $user_groups[[20]]$group
#> $user_groups[[20]]$group$id
#> [1] 1206
#> 
#> 
#> 
#> $user_groups[[21]]
#> $user_groups[[21]]$id
#> [1] 1994
#> 
#> $user_groups[[21]]$type
#> [1] "owner"
#> 
#> $user_groups[[21]]$security_policy
#> NULL
#> 
#> $user_groups[[21]]$group
#> $user_groups[[21]]$group$id
#> [1] 1207
#> 
#> 
#> 
#> $user_groups[[22]]
#> $user_groups[[22]]$id
#> [1] 1995
#> 
#> $user_groups[[22]]$type
#> [1] "owner"
#> 
#> $user_groups[[22]]$security_policy
#> NULL
#> 
#> $user_groups[[22]]$group
#> $user_groups[[22]]$group$id
#> [1] 1208
#> 
#> 
#> 
#> $user_groups[[23]]
#> $user_groups[[23]]$id
#> [1] 2000
#> 
#> $user_groups[[23]]$type
#> [1] "owner"
#> 
#> $user_groups[[23]]$security_policy
#> NULL
#> 
#> $user_groups[[23]]$group
#> $user_groups[[23]]$group$id
#> [1] 1213
#> 
#> 
#> 
#> $user_groups[[24]]
#> $user_groups[[24]]$id
#> [1] 2001
#> 
#> $user_groups[[24]]$type
#> [1] "owner"
#> 
#> $user_groups[[24]]$security_policy
#> NULL
#> 
#> $user_groups[[24]]$group
#> $user_groups[[24]]$group$id
#> [1] 1214
#> 
#> 
#> 
#> $user_groups[[25]]
#> $user_groups[[25]]$id
#> [1] 2009
#> 
#> $user_groups[[25]]$type
#> [1] "owner"
#> 
#> $user_groups[[25]]$security_policy
#> NULL
#> 
#> $user_groups[[25]]$group
#> $user_groups[[25]]$group$id
#> [1] 1220
#> 
#> 
#> 
#> $user_groups[[26]]
#> $user_groups[[26]]$id
#> [1] 2010
#> 
#> $user_groups[[26]]$type
#> [1] "owner"
#> 
#> $user_groups[[26]]$security_policy
#> NULL
#> 
#> $user_groups[[26]]$group
#> $user_groups[[26]]$group$id
#> [1] 1221
#> 
#> 
#> 
#> $user_groups[[27]]
#> $user_groups[[27]]$id
#> [1] 2011
#> 
#> $user_groups[[27]]$type
#> [1] "owner"
#> 
#> $user_groups[[27]]$security_policy
#> NULL
#> 
#> $user_groups[[27]]$group
#> $user_groups[[27]]$group$id
#> [1] 1222
#> 
#> 
#> 
#> 
#> $permissions
#> $permissions$view
#> [1] FALSE
#> 
#> $permissions$edit
#> [1] FALSE
#> 
#> $permissions$owner
#> [1] FALSE
#> 
#> 
#> $token
#> [1] "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJpYXQiOjE1NTQyMjI0NTcsImV4cCI6MTU1NDIyNjA1Nywicm9sZXMiOlsiUk9MRV9VU0VSIiwiUk9MRV9HUk9VUF8xMTgyIiwiUk9MRV9HUk9VUF9PV05FUl8xMTgyIiwiUk9MRV9HUk9VUF8xMTgzIiwiUk9MRV9HUk9VUF9PV05FUl8xMTgzIiwiUk9MRV9HUk9VUF8xMTg0IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg0IiwiUk9MRV9HUk9VUF8xMTg1IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg1IiwiUk9MRV9HUk9VUF8xMTg2IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg2IiwiUk9MRV9HUk9VUF8xMTg3IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg3IiwiUk9MRV9HUk9VUF8xMTg4IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg4IiwiUk9MRV9HUk9VUF8xMTg5IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg5IiwiUk9MRV9HUk9VUF8xMTkwIiwiUk9MRV9HUk9VUF9PV05FUl8xMTkwIiwiUk9MRV9HUk9VUF8xMTkxIiwiUk9MRV9HUk9VUF9PV05FUl8xMTkxIiwiUk9MRV9HUk9VUF8xMTkyIiwiUk9MRV9HUk9VUF9PV05FUl8xMTkyIiwiUk9MRV9HUk9VUF8xMTkzIiwiUk9MRV9HUk9VUF9PV05FUl8xMTkzIiwiUk9MRV9HUk9VUF8xMTk4IiwiUk9MRV9HUk9VUF9PV05FUl8xMTk4IiwiUk9MRV9HUk9VUF8xMjAwIiwiUk9MRV9HUk9VUF9PV05FUl8xMjAwIiwiUk9MRV9HUk9VUF8xMjAxIiwiUk9MRV9HUk9VUF9PV05FUl8xMjAxIiwiUk9MRV9HUk9VUF8xMjAyIiwiUk9MRV9HUk9VUF9PV05FUl8xMjAyIiwiUk9MRV9HUk9VUF8xMjAzIiwiUk9MRV9HUk9VUF9PV05FUl8xMjAzIiwiUk9MRV9HUk9VUF8xMjA0IiwiUk9MRV9HUk9VUF9PV05FUl8xMjA0IiwiUk9MRV9HUk9VUF8xMjA1IiwiUk9MRV9HUk9VUF9PV05FUl8xMjA1IiwiUk9MRV9HUk9VUF8xMjA2IiwiUk9MRV9HUk9VUF9PV05FUl8xMjA2IiwiUk9MRV9HUk9VUF8xMjA3IiwiUk9MRV9HUk9VUF9PV05FUl8xMjA3IiwiUk9MRV9HUk9VUF8xMjA4IiwiUk9MRV9HUk9VUF9PV05FUl8xMjA4IiwiUk9MRV9HUk9VUF8xMjEzIiwiUk9MRV9HUk9VUF9PV05FUl8xMjEzIiwiUk9MRV9HUk9VUF8xMjE0IiwiUk9MRV9HUk9VUF9PV05FUl8xMjE0IiwiUk9MRV9HUk9VUF8xMjIwIiwiUk9MRV9HUk9VUF9PV05FUl8xMjIwIiwiUk9MRV9HUk9VUF8xMjIxIiwiUk9MRV9HUk9VUF9PV05FUl8xMjIxIiwiUk9MRV9HUk9VUF8xMjIyIiwiUk9MRV9HUk9VUF9PV05FUl8xMjIyIl0sInVzZXJuYW1lIjoiZGVtb0BsYWJzdGVwLmNvbSIsInVzZXJfaWQiOjQ1MjR9.MbTyoXQEKX2ec27SpPBHYJg0CYBJCboQ7JpLJgF7YFLZ2GO90aDtTK5j1xqfenjAqtA1aVQuslagZfNtSKK9Vuav1o9Ewv3yAnhws1yPWiXQBYD4e3h5b5W0R1mr4b16sCUTOIbXXsCV-D8qQ1qN8ywRa-ib1mr-GuTEYmxPb6-1mfLvxtH45Y6b9QYVOZej-hWPhXy4AI7L8oLuoJJoZRDnCDdI8C1lZFLI24fWIZPD1kMkXxv8Uc19nIWF7MBF_MAnPpQvVBoucY6CZ49FXK7EPp5ZYByiw3AHosC3muhtDx2I-V6oBqKdRe7Mpvjxx_-zWaS66hZQzfa2JVp5SHlanJz8yHzj0id6dCdA8bSxYXp99wa7mF7ZzKuGIk12z5zIXGycxxEk3Q6GyhTiSpz_i8H9eTuV9XxrLHuc9w8HBT6M7rq_8lv_m2dUiI5qOXONCqXT4al_wbZxz1nSABwVirzJXKDh_an9XDpZPDhxiu7KeGgnc_sy91O0_zvrZ8b4JqhZgbQckF4OUU0OHGiO8W6V6AGOkYr8qENMAYigd9F24BAjsl259fa-uBH5pwWE1R3BGitVmGf28mOGKizKfFruh6S8yKrurf3wAm4HB5mFa8tX0Ub6FdzhqUY4tqDuSZEhkDmMkwfNLPelvacJM_1mQdlrbhyCPQefh9s"
#> 
#> $refresh_token
#> [1] "f4188377eda50b34cd71bcedc132571b122f107f9fbc19be3fcc29319171ddb18f175e51c5125eeb941cdfa49ec5c816cc09869829d267c339a07d87a8c9920f"
#> 
#> $is_new
#> NULL
#>