Create a Protocol on Labstep

This function allows you to create a protocol.

createProtocol(user, name)

Arguments

user
A labstep user object. Must contain an `api_key` field. Returned from `login` command

name
The name you want to give your protocol.

Value

Returns an `protocol` object

Examples

user <- login("demo@labstep.com","demopassword")
protocol <- createProtocol(user,'My First Protocol')
print(protocol)

#> $parent_protocol
#> NULL
#> 
#> $last_version
#> $last_version$all_experiment_count
#> [1] 0
#> 
#> $last_version$name
#> [1] "My First Protocol"
#> 
#> $last_version$experiment_count
#> [1] 0
#> 
#> $last_version$protocol_step_count
#> [1] 1
#> 
#> $last_version$protocol_table_count
#> [1] 0
#> 
#> $last_version$protocol_timer_count
#> [1] 0
#> 
#> $last_version$protocol_value_count
#> [1] 0
#> 
#> $last_version$children_protocol_collection_count
#> [1] 0
#> 
#> $last_version$version
#> [1] 0
#> 
#> $last_version$is_last_version
#> [1] TRUE
#> 
#> $last_version$id
#> [1] 16736
#> 
#> $last_version$changes
#> NULL
#> 
#> $last_version$additional_information
#> NULL
#> 
#> $last_version$content_state
#> $last_version$content_state$object
#> [1] "value"
#> 
#> $last_version$content_state$document
#> $last_version$content_state$document$data
#> list()
#> 
#> $last_version$content_state$document$nodes
#> $last_version$content_state$document$nodes[[1]]
#> $last_version$content_state$document$nodes[[1]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[1]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[1]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[1]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[1]]$nodes
#> list()
#> 
#> 
#> $last_version$content_state$document$nodes[[2]]
#> $last_version$content_state$document$nodes[[2]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[2]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[2]]$type
#> [1] "heading"
#> 
#> $last_version$content_state$document$nodes[[2]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[2]]$nodes
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Introduction"
#> 
#> $last_version$content_state$document$nodes[[2]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[3]]
#> $last_version$content_state$document$nodes[[3]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[3]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[3]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[3]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[3]]$nodes
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Get started by giving your protocol a name and by adding an introduction..."
#> 
#> $last_version$content_state$document$nodes[[3]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[4]]
#> $last_version$content_state$document$nodes[[4]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[4]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[4]]$type
#> [1] "heading"
#> 
#> $last_version$content_state$document$nodes[[4]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[4]]$nodes
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Materials"
#> 
#> $last_version$content_state$document$nodes[[4]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[5]]
#> $last_version$content_state$document$nodes[[5]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[5]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[5]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[5]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[5]]$nodes
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$text
#> [1] "List the required materials here."
#> 
#> $last_version$content_state$document$nodes[[5]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[6]]
#> $last_version$content_state$document$nodes[[6]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[6]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[6]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[6]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Create interactive elements by clicking the flask icon to allow you to track the inventory item and amount used each time you run the protocol."
#> 
#> $last_version$content_state$document$nodes[[6]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[7]]
#> $last_version$content_state$document$nodes[[7]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[7]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[7]]$type
#> [1] "heading"
#> 
#> $last_version$content_state$document$nodes[[7]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[7]]$nodes
#> $last_version$content_state$document$nodes[[7]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[7]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $last_version$content_state$document$nodes[[7]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[7]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[7]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $last_version$content_state$document$nodes[[7]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Procedure"
#> 
#> $last_version$content_state$document$nodes[[7]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[8]]
#> $last_version$content_state$document$nodes[[8]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[8]]$type
#> [1] "protocol_step"
#> 
#> $last_version$content_state$document$nodes[[8]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[8]]$data
#> $last_version$content_state$document$nodes[[8]]$data$id
#> [1] 91047
#> 
#> 
#> $last_version$content_state$document$nodes[[8]]$nodes
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$nodes
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$text
#> [1] "Detail your protocol procedure in steps like this. To add another step, click below 'Step 1' and select the plus icon on the left hand side."
#> 
#> $last_version$content_state$document$nodes[[8]]$nodes[[1]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[9]]
#> $last_version$content_state$document$nodes[[9]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[9]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[9]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[9]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[9]]$nodes
#> list()
#> 
#> 
#> $last_version$content_state$document$nodes[[10]]
#> $last_version$content_state$document$nodes[[10]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[10]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[10]]$type
#> [1] "heading"
#> 
#> $last_version$content_state$document$nodes[[10]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[10]]$nodes
#> $last_version$content_state$document$nodes[[10]]$nodes[[1]]
#> $last_version$content_state$document$nodes[[10]]$nodes[[1]]$object
#> [1] "text"
#> 
#> $last_version$content_state$document$nodes[[10]]$nodes[[1]]$leaves
#> $last_version$content_state$document$nodes[[10]]$nodes[[1]]$leaves[[1]]
#> $last_version$content_state$document$nodes[[10]]$nodes[[1]]$leaves[[1]]$object
#> [1] "leaf"
#> 
#> $last_version$content_state$document$nodes[[10]]$nodes[[1]]$leaves[[1]]$text
#> [1] "References"
#> 
#> $last_version$content_state$document$nodes[[10]]$nodes[[1]]$leaves[[1]]$marks
#> list()
#> 
#> 
#> 
#> 
#> 
#> 
#> $last_version$content_state$document$nodes[[11]]
#> $last_version$content_state$document$nodes[[11]]$object
#> [1] "block"
#> 
#> $last_version$content_state$document$nodes[[11]]$type
#> [1] "paragraph"
#> 
#> $last_version$content_state$document$nodes[[11]]$data
#> list()
#> 
#> $last_version$content_state$document$nodes[[11]]$isVoid
#> [1] FALSE
#> 
#> $last_version$content_state$document$nodes[[11]]$nodes
#> list()
#> 
#> 
#> 
#> $last_version$content_state$document$object
#> [1] "document"
#> 
#> 
#> 
#> $last_version$author
#> $last_version$author$id
#> [1] 4524
#> 
#> $last_version$author$username
#> [1] "demo@labstep.com"
#> 
#> $last_version$author$profile
#> $last_version$author$profile$id
#> [1] 5751
#> 
#> $last_version$author$profile$photo
#> NULL
#> 
#> 
#> 
#> $last_version$protocol_steps
#> $last_version$protocol_steps[[1]]
#> $last_version$protocol_steps[[1]]$id
#> [1] 91047
#> 
#> $last_version$protocol_steps[[1]]$files
#> list()
#> 
#> $last_version$protocol_steps[[1]]$protocol_tables
#> list()
#> 
#> $last_version$protocol_steps[[1]]$protocol_timers
#> list()
#> 
#> $last_version$protocol_steps[[1]]$protocol_values
#> list()
#> 
#> 
#> 
#> $last_version$protocol_tables
#> list()
#> 
#> $last_version$protocol_timers
#> list()
#> 
#> $last_version$protocol_values
#> list()
#> 
#> $last_version$files
#> list()
#> 
#> 
#> $id
#> [1] 13633
#> 
#> $created_at
#> [1] "2020-01-30T16:35:12+00:00"
#> 
#> $updated_at
#> [1] "2020-01-30T16:35:12+00:00"
#> 
#> $name
#> [1] "My First Protocol"
#> 
#> $deleted_at
#> NULL
#> 
#> $author
#> $author$id
#> [1] 4524
#> 
#> $author$username
#> [1] "demo@labstep.com"
#> 
#> $author$profile
#> $author$profile$id
#> [1] 5751
#> 
#> $author$profile$photo
#> NULL
#> 
#> 
#> 
#> $thread
#> $thread$all_comment_count
#> [1] 0
#> 
#> $thread$comment_count
#> [1] 0
#> 
#> $thread$id
#> [1] 673380
#> 
#> 
#> $perma_link
#> $perma_link$id
#> [1] 480263
#> 
#> $perma_link$token
#> [1] "b2e2d0eb-4c94-428b-bc3b-1619c0985a19"
#> 
#> 
#> $share_link
#> NULL
#> 
#> $permissions
#> $permissions$view
#> [1] TRUE
#> 
#> $permissions$edit
#> [1] TRUE
#> 
#> $permissions$owner
#> [1] TRUE
#> 
#> 
#> $owner
#> $owner$id
#> [1] 6867
#> 
#> $owner$name
#> [1] "Demo Labstep workspace"
#> 
#> $owner$created_at
#> [1] "2019-08-06T20:34:00+00:00"
#> 
#> $owner$deleted_at
#> NULL
#> 
#> $owner$description
#> NULL
#> 
#> $owner$profile
#> $owner$profile$id
#> [1] 19084
#> 
#> $owner$profile$photo
#> NULL
#> 
#> 
#> $owner$is_home
#> [1] TRUE
#> 
#>