Authenticate a Labstep user

This function allows you to retrieve the user details for your Labstep account. It also returns your unique API key which you need to use the other Labstep API functions.

login(username, password)

Arguments

username
Your Labstep username (usually the email address you signed up with)

password
Your Labstep password

Value

Returns a 'user' object.

Examples

user <- login("demo@labstep.com","demopassword")
print(user)

#> $id
#> [1] 4524
#> 
#> $username
#> [1] "demo@labstep.com"
#> 
#> $enabled
#> [1] TRUE
#> 
#> $groups
#> $groups[[1]]
#> $groups[[1]]$id
#> [1] 1182
#> 
#> 
#> $groups[[2]]
#> $groups[[2]]$id
#> [1] 1183
#> 
#> 
#> $groups[[3]]
#> $groups[[3]]$id
#> [1] 1184
#> 
#> 
#> $groups[[4]]
#> $groups[[4]]$id
#> [1] 1185
#> 
#> 
#> $groups[[5]]
#> $groups[[5]]$id
#> [1] 1186
#> 
#> 
#> $groups[[6]]
#> $groups[[6]]$id
#> [1] 1187
#> 
#> 
#> $groups[[7]]
#> $groups[[7]]$id
#> [1] 1188
#> 
#> 
#> $groups[[8]]
#> $groups[[8]]$id
#> [1] 1189
#> 
#> 
#> $groups[[9]]
#> $groups[[9]]$id
#> [1] 1190
#> 
#> 
#> $groups[[10]]
#> $groups[[10]]$id
#> [1] 1191
#> 
#> 
#> $groups[[11]]
#> $groups[[11]]$id
#> [1] 1192
#> 
#> 
#> $groups[[12]]
#> $groups[[12]]$id
#> [1] 1193
#> 
#> 
#> $groups[[13]]
#> $groups[[13]]$id
#> [1] 1198
#> 
#> 
#> $groups[[14]]
#> $groups[[14]]$id
#> [1] 1200
#> 
#> 
#> $groups[[15]]
#> $groups[[15]]$id
#> [1] 1201
#> 
#> 
#> $groups[[16]]
#> $groups[[16]]$id
#> [1] 1202
#> 
#> 
#> $groups[[17]]
#> $groups[[17]]$id
#> [1] 1203
#> 
#> 
#> $groups[[18]]
#> $groups[[18]]$id
#> [1] 1204
#> 
#> 
#> $groups[[19]]
#> $groups[[19]]$id
#> [1] 1205
#> 
#> 
#> 
#> $roles
#> $roles[[1]]
#> [1] "ROLE_USER"
#> 
#> $roles[[2]]
#> [1] "ROLE_GROUP_1182"
#> 
#> $roles[[3]]
#> [1] "ROLE_GROUP_OWNER_1182"
#> 
#> $roles[[4]]
#> [1] "ROLE_GROUP_1183"
#> 
#> $roles[[5]]
#> [1] "ROLE_GROUP_OWNER_1183"
#> 
#> $roles[[6]]
#> [1] "ROLE_GROUP_1184"
#> 
#> $roles[[7]]
#> [1] "ROLE_GROUP_OWNER_1184"
#> 
#> $roles[[8]]
#> [1] "ROLE_GROUP_1185"
#> 
#> $roles[[9]]
#> [1] "ROLE_GROUP_OWNER_1185"
#> 
#> $roles[[10]]
#> [1] "ROLE_GROUP_1186"
#> 
#> $roles[[11]]
#> [1] "ROLE_GROUP_OWNER_1186"
#> 
#> $roles[[12]]
#> [1] "ROLE_GROUP_1187"
#> 
#> $roles[[13]]
#> [1] "ROLE_GROUP_OWNER_1187"
#> 
#> $roles[[14]]
#> [1] "ROLE_GROUP_1188"
#> 
#> $roles[[15]]
#> [1] "ROLE_GROUP_OWNER_1188"
#> 
#> $roles[[16]]
#> [1] "ROLE_GROUP_1189"
#> 
#> $roles[[17]]
#> [1] "ROLE_GROUP_OWNER_1189"
#> 
#> $roles[[18]]
#> [1] "ROLE_GROUP_1190"
#> 
#> $roles[[19]]
#> [1] "ROLE_GROUP_OWNER_1190"
#> 
#> $roles[[20]]
#> [1] "ROLE_GROUP_1191"
#> 
#> $roles[[21]]
#> [1] "ROLE_GROUP_OWNER_1191"
#> 
#> $roles[[22]]
#> [1] "ROLE_GROUP_1192"
#> 
#> $roles[[23]]
#> [1] "ROLE_GROUP_OWNER_1192"
#> 
#> $roles[[24]]
#> [1] "ROLE_GROUP_1193"
#> 
#> $roles[[25]]
#> [1] "ROLE_GROUP_OWNER_1193"
#> 
#> $roles[[26]]
#> [1] "ROLE_GROUP_1198"
#> 
#> $roles[[27]]
#> [1] "ROLE_GROUP_OWNER_1198"
#> 
#> $roles[[28]]
#> [1] "ROLE_GROUP_1200"
#> 
#> $roles[[29]]
#> [1] "ROLE_GROUP_OWNER_1200"
#> 
#> $roles[[30]]
#> [1] "ROLE_GROUP_1201"
#> 
#> $roles[[31]]
#> [1] "ROLE_GROUP_OWNER_1201"
#> 
#> $roles[[32]]
#> [1] "ROLE_GROUP_1202"
#> 
#> $roles[[33]]
#> [1] "ROLE_GROUP_OWNER_1202"
#> 
#> $roles[[34]]
#> [1] "ROLE_GROUP_1203"
#> 
#> $roles[[35]]
#> [1] "ROLE_GROUP_OWNER_1203"
#> 
#> $roles[[36]]
#> [1] "ROLE_GROUP_1204"
#> 
#> $roles[[37]]
#> [1] "ROLE_GROUP_OWNER_1204"
#> 
#> $roles[[38]]
#> [1] "ROLE_GROUP_1205"
#> 
#> $roles[[39]]
#> [1] "ROLE_GROUP_OWNER_1205"
#> 
#> 
#> $group_count
#> [1] 19
#> 
#> $name
#> [1] "Demo Labstep"
#> 
#> $authenticated
#> NULL
#> 
#> $auth_provider
#> NULL
#> 
#> $created_at
#> [1] "2019-03-25T17:08:28+00:00"
#> 
#> $consented_at
#> NULL
#> 
#> $first_name
#> [1] "Demo"
#> 
#> $last_name
#> [1] "Labstep"
#> 
#> $api_key
#> [1] "NmQxZjUzNTZlYzZiZDk0M2NmZTc2YzBlOGY3NGM2ZDdlZWM0NWMzMg=="
#> 
#> $experience
#> [1] 0
#> 
#> $level
#> [1] 1
#> 
#> $profile
#> $profile$id
#> [1] 5751
#> 
#> $profile$about
#> NULL
#> 
#> $profile$photo
#> NULL
#> 
#> $profile$university
#> NULL
#> 
#> $profile$papers
#> list()
#> 
#> 
#> $protocol_collection_playlists
#> $protocol_collection_playlists[[1]]
#> $protocol_collection_playlists[[1]]$id
#> [1] 4431
#> 
#> $protocol_collection_playlists[[1]]$name
#> [1] "My protocol playlist"
#> 
#> 
#> 
#> $resource_lists
#> $resource_lists[[1]]
#> $resource_lists[[1]]$id
#> [1] 4431
#> 
#> 
#> 
#> $user_groups
#> $user_groups[[1]]
#> $user_groups[[1]]$id
#> [1] 1969
#> 
#> $user_groups[[1]]$type
#> [1] "owner"
#> 
#> $user_groups[[1]]$security_policy
#> NULL
#> 
#> $user_groups[[1]]$group
#> $user_groups[[1]]$group$id
#> [1] 1182
#> 
#> 
#> 
#> $user_groups[[2]]
#> $user_groups[[2]]$id
#> [1] 1970
#> 
#> $user_groups[[2]]$type
#> [1] "owner"
#> 
#> $user_groups[[2]]$security_policy
#> NULL
#> 
#> $user_groups[[2]]$group
#> $user_groups[[2]]$group$id
#> [1] 1183
#> 
#> 
#> 
#> $user_groups[[3]]
#> $user_groups[[3]]$id
#> [1] 1971
#> 
#> $user_groups[[3]]$type
#> [1] "owner"
#> 
#> $user_groups[[3]]$security_policy
#> NULL
#> 
#> $user_groups[[3]]$group
#> $user_groups[[3]]$group$id
#> [1] 1184
#> 
#> 
#> 
#> $user_groups[[4]]
#> $user_groups[[4]]$id
#> [1] 1972
#> 
#> $user_groups[[4]]$type
#> [1] "owner"
#> 
#> $user_groups[[4]]$security_policy
#> NULL
#> 
#> $user_groups[[4]]$group
#> $user_groups[[4]]$group$id
#> [1] 1185
#> 
#> 
#> 
#> $user_groups[[5]]
#> $user_groups[[5]]$id
#> [1] 1973
#> 
#> $user_groups[[5]]$type
#> [1] "owner"
#> 
#> $user_groups[[5]]$security_policy
#> NULL
#> 
#> $user_groups[[5]]$group
#> $user_groups[[5]]$group$id
#> [1] 1186
#> 
#> 
#> 
#> $user_groups[[6]]
#> $user_groups[[6]]$id
#> [1] 1974
#> 
#> $user_groups[[6]]$type
#> [1] "owner"
#> 
#> $user_groups[[6]]$security_policy
#> NULL
#> 
#> $user_groups[[6]]$group
#> $user_groups[[6]]$group$id
#> [1] 1187
#> 
#> 
#> 
#> $user_groups[[7]]
#> $user_groups[[7]]$id
#> [1] 1975
#> 
#> $user_groups[[7]]$type
#> [1] "owner"
#> 
#> $user_groups[[7]]$security_policy
#> NULL
#> 
#> $user_groups[[7]]$group
#> $user_groups[[7]]$group$id
#> [1] 1188
#> 
#> 
#> 
#> $user_groups[[8]]
#> $user_groups[[8]]$id
#> [1] 1976
#> 
#> $user_groups[[8]]$type
#> [1] "owner"
#> 
#> $user_groups[[8]]$security_policy
#> NULL
#> 
#> $user_groups[[8]]$group
#> $user_groups[[8]]$group$id
#> [1] 1189
#> 
#> 
#> 
#> $user_groups[[9]]
#> $user_groups[[9]]$id
#> [1] 1977
#> 
#> $user_groups[[9]]$type
#> [1] "owner"
#> 
#> $user_groups[[9]]$security_policy
#> NULL
#> 
#> $user_groups[[9]]$group
#> $user_groups[[9]]$group$id
#> [1] 1190
#> 
#> 
#> 
#> $user_groups[[10]]
#> $user_groups[[10]]$id
#> [1] 1978
#> 
#> $user_groups[[10]]$type
#> [1] "owner"
#> 
#> $user_groups[[10]]$security_policy
#> NULL
#> 
#> $user_groups[[10]]$group
#> $user_groups[[10]]$group$id
#> [1] 1191
#> 
#> 
#> 
#> $user_groups[[11]]
#> $user_groups[[11]]$id
#> [1] 1979
#> 
#> $user_groups[[11]]$type
#> [1] "owner"
#> 
#> $user_groups[[11]]$security_policy
#> NULL
#> 
#> $user_groups[[11]]$group
#> $user_groups[[11]]$group$id
#> [1] 1192
#> 
#> 
#> 
#> $user_groups[[12]]
#> $user_groups[[12]]$id
#> [1] 1980
#> 
#> $user_groups[[12]]$type
#> [1] "owner"
#> 
#> $user_groups[[12]]$security_policy
#> NULL
#> 
#> $user_groups[[12]]$group
#> $user_groups[[12]]$group$id
#> [1] 1193
#> 
#> 
#> 
#> $user_groups[[13]]
#> $user_groups[[13]]$id
#> [1] 1985
#> 
#> $user_groups[[13]]$type
#> [1] "owner"
#> 
#> $user_groups[[13]]$security_policy
#> NULL
#> 
#> $user_groups[[13]]$group
#> $user_groups[[13]]$group$id
#> [1] 1198
#> 
#> 
#> 
#> $user_groups[[14]]
#> $user_groups[[14]]$id
#> [1] 1987
#> 
#> $user_groups[[14]]$type
#> [1] "owner"
#> 
#> $user_groups[[14]]$security_policy
#> NULL
#> 
#> $user_groups[[14]]$group
#> $user_groups[[14]]$group$id
#> [1] 1200
#> 
#> 
#> 
#> $user_groups[[15]]
#> $user_groups[[15]]$id
#> [1] 1988
#> 
#> $user_groups[[15]]$type
#> [1] "owner"
#> 
#> $user_groups[[15]]$security_policy
#> NULL
#> 
#> $user_groups[[15]]$group
#> $user_groups[[15]]$group$id
#> [1] 1201
#> 
#> 
#> 
#> $user_groups[[16]]
#> $user_groups[[16]]$id
#> [1] 1989
#> 
#> $user_groups[[16]]$type
#> [1] "owner"
#> 
#> $user_groups[[16]]$security_policy
#> NULL
#> 
#> $user_groups[[16]]$group
#> $user_groups[[16]]$group$id
#> [1] 1202
#> 
#> 
#> 
#> $user_groups[[17]]
#> $user_groups[[17]]$id
#> [1] 1990
#> 
#> $user_groups[[17]]$type
#> [1] "owner"
#> 
#> $user_groups[[17]]$security_policy
#> NULL
#> 
#> $user_groups[[17]]$group
#> $user_groups[[17]]$group$id
#> [1] 1203
#> 
#> 
#> 
#> $user_groups[[18]]
#> $user_groups[[18]]$id
#> [1] 1991
#> 
#> $user_groups[[18]]$type
#> [1] "owner"
#> 
#> $user_groups[[18]]$security_policy
#> NULL
#> 
#> $user_groups[[18]]$group
#> $user_groups[[18]]$group$id
#> [1] 1204
#> 
#> 
#> 
#> $user_groups[[19]]
#> $user_groups[[19]]$id
#> [1] 1992
#> 
#> $user_groups[[19]]$type
#> [1] "owner"
#> 
#> $user_groups[[19]]$security_policy
#> NULL
#> 
#> $user_groups[[19]]$group
#> $user_groups[[19]]$group$id
#> [1] 1205
#> 
#> 
#> 
#> 
#> $permissions
#> $permissions$view
#> [1] FALSE
#> 
#> $permissions$edit
#> [1] FALSE
#> 
#> $permissions$owner
#> [1] FALSE
#> 
#> 
#> $token
#> [1] "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJpYXQiOjE1NTQxMzc1MTYsImV4cCI6MTU1NDE0MTExNiwicm9sZXMiOlsiUk9MRV9VU0VSIiwiUk9MRV9HUk9VUF8xMTgyIiwiUk9MRV9HUk9VUF9PV05FUl8xMTgyIiwiUk9MRV9HUk9VUF8xMTgzIiwiUk9MRV9HUk9VUF9PV05FUl8xMTgzIiwiUk9MRV9HUk9VUF8xMTg0IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg0IiwiUk9MRV9HUk9VUF8xMTg1IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg1IiwiUk9MRV9HUk9VUF8xMTg2IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg2IiwiUk9MRV9HUk9VUF8xMTg3IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg3IiwiUk9MRV9HUk9VUF8xMTg4IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg4IiwiUk9MRV9HUk9VUF8xMTg5IiwiUk9MRV9HUk9VUF9PV05FUl8xMTg5IiwiUk9MRV9HUk9VUF8xMTkwIiwiUk9MRV9HUk9VUF9PV05FUl8xMTkwIiwiUk9MRV9HUk9VUF8xMTkxIiwiUk9MRV9HUk9VUF9PV05FUl8xMTkxIiwiUk9MRV9HUk9VUF8xMTkyIiwiUk9MRV9HUk9VUF9PV05FUl8xMTkyIiwiUk9MRV9HUk9VUF8xMTkzIiwiUk9MRV9HUk9VUF9PV05FUl8xMTkzIiwiUk9MRV9HUk9VUF8xMTk4IiwiUk9MRV9HUk9VUF9PV05FUl8xMTk4IiwiUk9MRV9HUk9VUF8xMjAwIiwiUk9MRV9HUk9VUF9PV05FUl8xMjAwIiwiUk9MRV9HUk9VUF8xMjAxIiwiUk9MRV9HUk9VUF9PV05FUl8xMjAxIiwiUk9MRV9HUk9VUF8xMjAyIiwiUk9MRV9HUk9VUF9PV05FUl8xMjAyIiwiUk9MRV9HUk9VUF8xMjAzIiwiUk9MRV9HUk9VUF9PV05FUl8xMjAzIiwiUk9MRV9HUk9VUF8xMjA0IiwiUk9MRV9HUk9VUF9PV05FUl8xMjA0IiwiUk9MRV9HUk9VUF8xMjA1IiwiUk9MRV9HUk9VUF9PV05FUl8xMjA1Il0sInVzZXJuYW1lIjoiZGVtb0BsYWJzdGVwLmNvbSIsInVzZXJfaWQiOjQ1MjR9.cpngB6y1dtfww9PPtWzBdY31xDYnhot7wukMhNIMjhffwBrXQD8XkWHLf7HexeHJNPccgHBs3g5fHcSqdAAXRwQqAujloa2mXcK07zKo1YKOwVTe2kc3-XnbZe2cpvC5EXx5imHi-DW3rqUXVzFfvmtYenz79UlCxMMJ2w-SF3wXQ3S8pIPRsWSrI0Og1u8m4rdXMyThDqb0Cje-ZH4_LstZUSdOgd8XQwIk-OcqbiN4oYIALPaLhyZXhuKXhYTsIVuo94M-IsstpRbbZvv4Wx8_6C1Dz7nwtCAC3m7uBtiZFQawe_MZr5APv5zE14CM_E6wdi04iXV3d7VIs_Qqurb3nwPCP-NobSuAR_7PSK20PKjIXMZo3wwZH6X54yJlpDdbI9TjQAXIiMe9vSJmVUB0Agi9RousNzXI990Tdy2hnMceC1YCUi_b63PBRLGAt0TjY2y4dQaN-iL5IrOEHULY-ABmnRxoePkw0naoks2mki2TKazT1zCtiRGxui6QmSqTqLQix5NE4cQeb0fl2O9W1V7sVDE-Zrnbc4HtXRYWgPqYyxToerPjen5uSru-WwuDWEPfTPgE8PCUbvIW_ta19fo3UK-kyICMppd2urD4XGLWMJ_n0eS7G_6DlsF2kevYQGkzaC7-9my-R33YtVGYm7WNzQGyRUy_o1OaTvQ"
#> 
#> $refresh_token
#> [1] "c67ceaa710214ded666f24cc11557a613807b0dea784b8a9de4cedd8385dd8b5cbf2e2bc6719a02de209682e232a8982e271f46240104f32d92ec2e4ae7a6c19"
#> 
#> $is_new
#> NULL
#>